live/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
idna==2.9
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.18.4
pandas==1.0.3
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
squarify==0.4.3
urllib3==1.25.9
